requirements.txt
# Google API for Email
google-api-python-client
google-auth-oauthlib
google-auth-httplib2

# LLM Interaction
openai # For LM Studio's OpenAI-compatible server
google-cloud-aiplatform

# Database
SQLAlchemy
chromadb

# Background Jobs
APScheduler

# Data validation (implicitly used by FastAPI, good to have explicit)
pydantic

# Web UI (Optional for now)
fastapi
uvicorn

# Packaging (Optional for now)
PyInstaller

# --- DEPENDENCY CONFLICT RESOLUTION ---
# Pin protobuf to a version compatible with both chromadb's dependencies
# and the google-cloud libraries to resolve the "Descriptors cannot be created directly" error.
protobuf==3.20.3